pyproject.toml
[tool.poetry]
name = "Akane"
version = "0.2.0"
description = "A general purpose and fun Discord bot."
authors = ["AbstractUmbra <Umbra@AbstractUmbra.xyz>"]
license = "MPL 2.0"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = ">=1.7.0"
discord-ext-menus = {git = "https://github.com/Rapptz/discord-ext-menus", rev = "master"}
jishaku = {git = "https://github.com/Gorialis/jishaku", rev = "master"}
asyncpg = "*"
googletrans = "3.1.0a0"
psutil = "*"
click = "*"
parsedatetime = "*"
lru_dict = "*"
lxml = "*"
pytz = "^2021.1"
python-dateutil = "*"
fuzzywuzzy = "*"
"mystbin.py" = "*"
feedparser = "*"
CurrencyConverter = "*"
"discord.py-stubs" = "*"
nhentaio = "*"
pykakasi = "*"
beautifulsoup4 = "*"
Pillow = "*"
pygit2 = "^1.5.0"

[tool.poetry.dev-dependencies]
black = "*"
flake8 = "*"
isort = "^5.8.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
